gemma-training/requirements.txt
# Gemma QLoRA Local Dependencies (T4 GPU + CUDA 12.4)
# Normalized for JAX-native training (No TensorFlow)
# ==============================================================================

# JAX and CUDA 12 support
jax[cuda12]
nvidia-cuda-runtime-cu12
nvidia-cuda-cupti-cu12
nvidia-cudnn-cu12
nvidia-cublas-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12

flax>=0.11.1
numpy>=2.1.3,<2.2.0
ml-dtypes>=0.5.0
optax>=0.2.7
orbax-checkpoint>=0.11.33
numba>=0.60.0,<0.62.0

# JAX-based fine-tuning frameworks
qwix==0.1.5
google-tunix==0.1.6

# Financial Data & LLM Utilities
clu
ml_collections
datasets
transformers
huggingface_hub
pymupdf
requests
tqdm
pandas
tabulate
python-dotenv

# Jupyter environment
ipykernel
jupyterlab
importlib_resources
